Extract MaybeCompressedDataReader from MaybeCompressedFileReader
